refactor: Simplify client-side ingredient and step creation for a recipe
Instead of trying to store the same data in server and in client via
HTML attributes, remove the overhead of moving the entered ingredient
data to a table and simply keep the entered data as-is in the inputs.
Change the event handlers so that they only add a new row of inputs
needed for ingredient and step data.